feat(F0): Использовано дерево вместо стандартного std::map и добавлено безопасное удаление в само дерево, а также добавлено удаление с параметром итератор, так как было использовано в ргр по ТП
