+++ mihalchenko.vladislav/F0/details.hpp
@@ -3,13 +3,13 @@
 #include <map>
 #include <string>
 #include <iostream>
+#include <AVLtree.hpp>
 
 namespace mihalchenko
 {
-  using pair_t = std::pair< std::string, size_t >;
   using dictElement_t = std::pair< std::string, size_t >;
-  using dict_t = std::map< std::string, size_t >;
-  using mapOfDicts_t = std::map< std::string, std::map< std::string, size_t > >;
+  using dict_t = AVLTree< std::string, size_t >;
+  using mapOfDicts_t = AVLTree< std::string, dict_t >;
 }
 
 #endif
