+++ mihalchenko.vladislav/F0/helpFunctions.hpp
@@ -8,8 +8,8 @@ namespace mihalchenko
   std::string getDictName(std::string &str);
   void sortByFreq(mapOfDicts_t &mapOfDictionaries, const std::string &nameOfDict, std::istream &is, std::ostream &out);
   void sortByAlph(mapOfDicts_t &mapOfDictionaries, const std::string &nameOfDict, std::istream &is, std::ostream &out);
-  bool compareElemByFreq(const pair_t &iterOfElem1, const pair_t &iterOfElem2);
-  bool compareElemByAlph(const pair_t &iterOfElem1, const pair_t &iterOfElem2);
+  bool compareElemByFreq(const dictElement_t &iterOfElem1, const dictElement_t &iterOfElem2);
+  bool compareElemByAlph(const dictElement_t &iterOfElem1, const dictElement_t &iterOfElem2);
   std::ostream &printInvalidCommand(std::ostream &out);
   std::ostream &printErrorMessage(std::ostream &out);
   std::ostream &printWrongInput(std::ostream &out);
